+++ samples/ansicolors/src/main/kotlin/com/github/ajalt/clikt/samples/ansicolors/main.kt
@@ -4,7 +4,7 @@ import com.github.ajalt.clikt.core.NoRunCliktCommand
 import com.github.ajalt.clikt.core.context
 import com.github.ajalt.clikt.core.subcommands
 import com.github.ajalt.clikt.output.HelpFormatter
-import com.github.ajalt.clikt.output.PlaintextHelpFormatter
+import com.github.ajalt.clikt.output.CliktHelpFormatter
 import com.github.ajalt.clikt.parameters.arguments.argument
 import com.github.ajalt.clikt.parameters.arguments.multiple
 import com.github.ajalt.clikt.parameters.options.flag
@@ -13,7 +13,7 @@ import com.github.ajalt.mordant.TermColors
 
 private val tc = TermColors(TermColors.Level.TRUECOLOR)
 
-class ColorHelpFormatter : PlaintextHelpFormatter(usageTitle = (tc.bold + tc.underline)("Usage:")) {
+class ColorHelpFormatter : CliktHelpFormatter(usageTitle = (tc.bold + tc.underline)("Usage:")) {
     override fun renderTag(tag: String, value: String) = tc.green(super.renderTag(tag, value))
     override fun renderOptionName(name: String) = tc.yellow(super.renderOptionName(name))
     override fun renderArgumentName(name: String) = tc.yellow(super.renderArgumentName(name))