+++ samples/helpformat/src/main/kotlin/com/github/ajalt/clikt/samples/helpformat/main.kt
@@ -3,14 +3,13 @@ package com.github.ajalt.clikt.samples.helpformat
 import com.github.ajalt.clikt.core.CliktCommand
 import com.github.ajalt.clikt.core.context
 import com.github.ajalt.clikt.output.HelpFormatter
-import com.github.ajalt.clikt.output.PlaintextHelpFormatter
-import com.github.ajalt.clikt.output.TermUi
+import com.github.ajalt.clikt.output.CliktHelpFormatter
 import com.github.ajalt.clikt.parameters.arguments.argument
 import com.github.ajalt.clikt.parameters.arguments.multiple
 import com.github.ajalt.clikt.parameters.options.flag
 import com.github.ajalt.clikt.parameters.options.option
 
-class ArgparseHelpFormatter: PlaintextHelpFormatter(
+class ArgparseHelpFormatter: CliktHelpFormatter(
         usageTitle = "usage:",
         optionsTitle = "optional arguments:",
         argumentsTitle = "positional arguments:") {